Fix runtime type checking for keyboard layout
