--- src/lib/keyboardlayouts.ts
@@ -1,352 +1,0 @@
-/**
- * Key codes for printable keys for [[keylayoutforce]], lower and upper register.
- * See https://developer.mozilla.org/en-US/docs/Web/API/UI_Events/Keyboard_event_code_values
- * These maps are assigned  `:setkeyboardlayout` command,
- * but keylayoutforcemapping can also be changed manually with `:set`.
- *
- * If your layout is missing here, you can contribute it with the help of [this](https://gistpreview.github.io/?324119c773fac31651f6422087b36804)
- * tool.
- */
-export const keyboardlayouts = {
-    qwerty: {
-        KeyA: ["a", "A"],
-        KeyB: ["b", "B"],
-        KeyC: ["c", "C"],
-        KeyD: ["d", "D"],
-        KeyE: ["e", "E"],
-        KeyF: ["f", "F"],
-        KeyG: ["g", "G"],
-        KeyH: ["h", "H"],
-        KeyI: ["i", "I"],
-        KeyJ: ["j", "J"],
-        KeyK: ["k", "K"],
-        KeyL: ["l", "L"],
-        KeyM: ["m", "M"],
-        KeyN: ["n", "N"],
-        KeyO: ["o", "O"],
-        KeyP: ["p", "P"],
-        KeyQ: ["q", "Q"],
-        KeyR: ["r", "R"],
-        KeyS: ["s", "S"],
-        KeyT: ["t", "T"],
-        KeyU: ["u", "U"],
-        KeyV: ["v", "V"],
-        KeyW: ["w", "W"],
-        KeyX: ["x", "X"],
-        KeyY: ["y", "Y"],
-        KeyZ: ["z", "Z"],
-        Digit0: ["0", ")"],
-        Digit1: ["1", "!"],
-        Digit2: ["2", "@"],
-        Digit3: ["3", "#"],
-        Digit4: ["4", "$"],
-        Digit5: ["5", "%"],
-        Digit6: ["6", "^"],
-        Digit7: ["7", "&"],
-        Digit8: ["8", "*"],
-        Digit9: ["9", "("],
-        Equal: ["=", "+"],
-        Backquote: ["`", "~"],
-        Backslash: ["\\", "|"],
-        Period: [".", ">"],
-        Comma: [",", "<"],
-        Semicolon: [";", ":"],
-        Slash: ["/", "?"],
-        BracketLeft: ["[", "{"],
-        BracketRight: ["]", "}"],
-        Quote: ["'", '"'],
-        Minus: ["-", "_"],
-    },
-    azerty: {
-        Backquote: ["²", "²"],
-        Digit1: ["&", "1"],
-        Digit2: ["é", "2"],
-        Digit3: ['"', "3"],
-        Digit4: ["'", "4"],
-        Digit5: ["(", "5"],
-        Digit6: ["-", "6"],
-        Digit7: ["è", "7"],
-        Digit8: ["_", "8"],
-        Digit9: ["ç", "9"],
-        Digit0: ["à", "0"],
-        Minus: [")", "°"],
-        Equal: ["=", "+"],
-        KeyQ: ["a", "A"],
-        KeyW: ["z", "Z"],
-        KeyE: ["e", "E"],
-        KeyR: ["r", "R"],
-        KeyT: ["t", "T"],
-        KeyY: ["y", "Y"],
-        KeyU: ["u", "U"],
-        KeyI: ["i", "I"],
-        KeyO: ["o", "O"],
-        KeyP: ["p", "P"],
-        BracketRight: ["$", "£"],
-        Backslash: ["*", "µ"],
-        KeyA: ["q", "Q"],
-        KeyS: ["s", "S"],
-        KeyD: ["d", "D"],
-        KeyF: ["f", "F"],
-        KeyG: ["g", "G"],
-        KeyH: ["h", "H"],
-        KeyJ: ["j", "J"],
-        KeyK: ["k", "K"],
-        KeyL: ["l", "L"],
-        Semicolon: ["m", "M"],
-        Quote: ["ù", "%"],
-        KeyZ: ["w", "W"],
-        KeyX: ["x", "X"],
-        KeyC: ["c", "C"],
-        KeyV: ["v", "V"],
-        KeyB: ["b", "B"],
-        KeyN: ["n", "N"],
-        KeyM: [",", "?"],
-        Comma: [";", "."],
-        Period: [":", "/"],
-        Slash: ["!", "§"],
-    },
-    german: {
-        Digit1: ["1", "!"],
-        Digit2: ["2", '"'],
-        Digit3: ["3", "§"],
-        Digit4: ["4", "$"],
-        Digit5: ["5", "%"],
-        Digit6: ["6", "&"],
-        Digit7: ["7", "/"],
-        Digit8: ["8", "("],
-        Digit9: ["9", ")"],
-        Digit0: ["0", "="],
-        Minus: ["ß", "?"],
-        KeyQ: ["q", "Q"],
-        KeyW: ["w", "W"],
-        KeyE: ["e", "E"],
-        KeyR: ["r", "R"],
-        KeyT: ["t", "T"],
-        KeyY: ["z", "Z"],
-        KeyU: ["u", "U"],
-        KeyI: ["i", "I"],
-        KeyO: ["o", "O"],
-        KeyP: ["p", "P"],
-        BracketLeft: ["ü", "Ü"],
-        BracketRight: ["+", "*"],
-        Backslash: ["#", "'"],
-        KeyA: ["a", "A"],
-        KeyS: ["s", "S"],
-        KeyD: ["d", "D"],
-        KeyF: ["f", "F"],
-        KeyG: ["g", "G"],
-        KeyH: ["h", "H"],
-        KeyJ: ["j", "J"],
-        KeyK: ["k", "K"],
-        KeyL: ["l", "L"],
-        Semicolon: ["ö", "Ö"],
-        Quote: ["ä", "Ä"],
-        KeyZ: ["y", "Y"],
-        KeyX: ["x", "X"],
-        KeyC: ["c", "C"],
-        KeyV: ["v", "V"],
-        KeyB: ["b", "B"],
-        KeyN: ["n", "N"],
-        KeyM: ["m", "M"],
-        Comma: [",", ";"],
-        Period: [".", ":"],
-        Slash: ["-", "_"],
-        Backquote: ["", "°"],
-    },
-    dvorak: {
-        Backquote: ["`", "~"],
-        Digit1: ["1", "!"],
-        Digit2: ["2", "@"],
-        Digit3: ["3", "#"],
-        Digit4: ["4", "$"],
-        Digit5: ["5", "%"],
-        Digit6: ["6", "^"],
-        Digit7: ["7", "&"],
-        Digit8: ["8", "*"],
-        Digit9: ["9", "("],
-        Digit0: ["0", ")"],
-        Minus: ["[", "{"],
-        Equal: ["]", "}"],
-        KeyQ: ["'", '"'],
-        KeyW: [",", "<"],
-        KeyE: [".", ">"],
-        KeyR: ["p", "P"],
-        KeyT: ["y", "Y"],
-        KeyY: ["f", "F"],
-        KeyU: ["g", "G"],
-        KeyI: ["c", "C"],
-        KeyO: ["r", "R"],
-        KeyP: ["l", "L"],
-        BracketLeft: ["/", "?"],
-        BracketRight: ["=", "+"],
-        Backslash: ["\\", "|"],
-        KeyA: ["a", "A"],
-        KeyS: ["o", "O"],
-        KeyD: ["e", "E"],
-        KeyF: ["u", "U"],
-        KeyG: ["i", "I"],
-        KeyH: ["d", "D"],
-        KeyJ: ["h", "H"],
-        KeyK: ["t", "T"],
-        KeyL: ["n", "N"],
-        Semicolon: ["s", "S"],
-        Quote: ["-", "_"],
-        KeyZ: [";", ":"],
-        KeyX: ["q", "Q"],
-        KeyC: ["j", "J"],
-        KeyV: ["k", "K"],
-        KeyB: ["x", "X"],
-        KeyN: ["b", "B"],
-        KeyM: ["m", "M"],
-        Comma: ["w", "W"],
-        Period: ["v", "V"],
-        Slash: ["z", "Z"],
-    },
-    uk: {
-        Digit1: ["1", "!"],
-        Digit2: ["2", '"'],
-        Digit3: ["3", "£"],
-        Digit4: ["4", "$"],
-        Digit5: ["5", "%"],
-        Digit6: ["6", "^"],
-        Digit7: ["7", "&"],
-        Digit8: ["8", "*"],
-        Digit9: ["9", "("],
-        Digit0: ["0", ")"],
-        Minus: ["-", "_"],
-        Equal: ["=", "+"],
-        KeyQ: ["q", "Q"],
-        KeyW: ["w", "W"],
-        KeyE: ["e", "E"],
-        KeyR: ["r", "R"],
-        KeyT: ["t", "T"],
-        KeyY: ["y", "Y"],
-        KeyU: ["u", "U"],
-        KeyI: ["i", "I"],
-        KeyO: ["o", "O"],
-        KeyP: ["p", "P"],
-        BracketLeft: ["[", "{"],
-        KeyK: ["k", "K"],
-        BracketRight: ["]", "}"],
-        KeyA: ["a", "A"],
-        KeyS: ["s", "S"],
-        KeyD: ["d", "D"],
-        KeyF: ["f", "F"],
-        KeyG: ["g", "G"],
-        KeyH: ["h", "H"],
-        KeyJ: ["j", "J"],
-        Semicolon: [";", ":"],
-        Quote: ["'", "@"],
-        Backslash: ["#", "~"],
-        IntlBackslash: ["\\", "|"],
-        KeyZ: ["z", "Z"],
-        KeyX: ["x", "X"],
-        KeyC: ["c", "C"],
-        KeyV: ["v", "V"],
-        KeyB: ["b", "B"],
-        KeyN: ["n", "N"],
-        KeyM: ["m", "M"],
-        Period: [".", ">"],
-        Slash: ["/", "?"],
-        Backquote: ["`", "¬"],
-        KeyL: ["l", "L"],
-        Comma: [",", "<"],
-    },
-    ca: {
-        Backquote: ["#", "|"],
-        Digit1: ["1", "!"],
-        Digit2: ["2", '"'],
-        Digit3: ["3", "/"],
-        Digit4: ["4", "$"],
-        Digit5: ["5", "%"],
-        Digit6: ["6", "?"],
-        Digit7: ["7", "&"],
-        Digit8: ["8", "*"],
-        Digit9: ["9", "("],
-        Digit0: ["0", ")"],
-        Minus: ["-", "_"],
-        Equal: ["=", "+"],
-        KeyQ: ["q", "Q"],
-        KeyW: ["w", "W"],
-        KeyE: ["e", "E"],
-        KeyR: ["r", "R"],
-        KeyT: ["t", "T"],
-        KeyY: ["y", "Y"],
-        KeyU: ["u", "U"],
-        KeyI: ["i", "I"],
-        KeyO: ["o", "O"],
-        KeyP: ["p", "P"],
-        KeyA: ["a", "A"],
-        KeyS: ["s", "S"],
-        KeyD: ["d", "D"],
-        KeyF: ["f", "F"],
-        KeyG: ["g", "G"],
-        KeyH: ["h", "H"],
-        KeyJ: ["j", "J"],
-        KeyK: ["k", "K"],
-        KeyL: ["l", "L"],
-        Semicolon: [";", ":"],
-        Backslash: ["<", ">"],
-        IntlBackslash: ["«", "»"],
-        KeyZ: ["z", "Z"],
-        KeyX: ["x", "X"],
-        KeyC: ["c", "C"],
-        KeyV: ["v", "V"],
-        KeyB: ["b", "B"],
-        KeyN: ["n", "N"],
-        KeyM: ["m", "M"],
-        Comma: [",", "'"],
-        Period: [".", "."],
-        Slash: ["é", "É"],
-    },
-    bepo: {
-        Backquote: ["$", "#"],
-        Digit1: ['"', "1"],
-        Digit2: ["«", "2"],
-        Digit3: ["»", "3"],
-        Digit4: ["(", "4"],
-        Digit5: [")", "5"],
-        Digit6: ["@", "6"],
-        Digit7: ["+", "7"],
-        Digit8: ["-", "8"],
-        Digit9: ["/", "9"],
-        Digit0: ["*", "0"],
-        Minus: ["=", "°"],
-        Equal: ["%", "`"],
-        KeyQ: ["b", "B"],
-        KeyW: ["é", "É"],
-        KeyE: ["p", "P"],
-        KeyR: ["o", "O"],
-        KeyT: ["è", "È"],
-        KeyU: ["v", "V"],
-        KeyI: ["d", "D"],
-        KeyO: ["l", "L"],
-        KeyP: ["j", "J"],
-        BracketLeft: ["z", "Z"],
-        BracketRight: ["w", "W"],
-        KeyA: ["a", "A"],
-        KeyS: ["u", "U"],
-        KeyD: ["i", "I"],
-        KeyF: ["e", "E"],
-        KeyG: [",", ";"],
-        KeyH: ["c", "C"],
-        KeyJ: ["t", "T"],
-        KeyK: ["s", "S"],
-        KeyL: ["r", "R"],
-        Semicolon: ["n", "N"],
-        Quote: ["m", "M"],
-        Backslash: ["ç", "Ç"],
-        IntlBackslash: ["ê", "Ê"],
-        KeyZ: ["à", "À"],
-        KeyX: ["y", "Y"],
-        KeyC: ["x", "X"],
-        KeyV: [".", ":"],
-        KeyB: ["k", "K"],
-        KeyN: ["'", "?"],
-        KeyM: ["q", "Q"],
-        Comma: ["g", "G"],
-        Period: ["h", "H"],
-        Slash: ["f", "F"],
-        KeyY: ["", "!"],
-    },
-}
